All current visual studio files
The "Neural Network" solution contains what I believe to be a fully functioning fully connected neural network, and possibly also a working convolutional neural network. The textToPictures solution converts each line of the text (.txt) files in \Neural network\Numbers\ into bitmap (.bmp) files in the subfolder corresponding with the number in the image, and a file name corresponding with the line of the text file it is stored in.