Backpropagation, and Images
Added a project to use backpropagation to improve neural net results,
rather than a genetic algorithm.
Removed the textToPictures class, and the text to go along with it.
Instead, took the output files of that, and put them in a zip file.
Before running any of the projects, this must be unzipped.